Add Hono as an option (start:standalone)
